Initial Project Setup / Employee Class
Initial project set up and the initial version of the Employee Class. Need to create Unit tests and pass them before I push the tasks to complete on projects page. Tried setting this on a new branch but wasn't able to. I think this is due to the repository being empty.