Updated tests for Employee class
Just quickly updated the tests to match the change from int to Strings for IDs. Also wrote some empty tests for the getAccess() method. This method will return a list of rooms, buildings ,and suites the employee has access too. Will add more to the tests once the building, suite, and room classes are complete.